fix: add ::integer casts to all COUNT(*)/SUM() aggregates in routes and services
PostgreSQL's pg driver serialises COUNT(*) and SUM() as strings.
Without explicit casts, arithmetic (e.g. division, comparisons) silently
operates on strings instead of numbers.

Files fixed:
- routes/metrics.ts – sample_count
- routes/monitoring.ts – insights pagination count
- routes/remediation.ts – total + pendingCount
- services/feedback-store.ts – list count, stats (total/positive/negative/pending_count), negative-count
- services/ebpf-coverage.ts – getCoverageSummary counts
- services/llm-trace-store.ts – total_queries, total_tokens, model breakdown count/tokens
- services/llm-tools.ts – call_count in slowest-endpoints query
- services/pcap-store.ts – getCapturesCount (both variants)
- services/trace-store.ts – span_count, callCount (nodes + edges)

All 1967 tests still pass.

+++ backend/src/routes/metrics.ts
@@ -250,7 +250,7 @@ export async function metricsRoutes(fastify: FastifyInstance) {
         AVG(value) as avg_value,
         MIN(value) as min_value,
         MAX(value) as max_value,
-        COUNT(*) as sample_count
+        COUNT(*)::integer as sample_count
       FROM metrics
       WHERE container_id = $1 AND timestamp >= $2
       GROUP BY metric_type
