+++ backend/src/services/llm-trace-store.ts
@@ -76,8 +76,8 @@ export async function getLlmStats(hoursBack: number = 24): Promise<LlmStats> {
     error_rate: number;
   }>(`
     SELECT
-      COUNT(*) as total_queries,
-      COALESCE(SUM(total_tokens), 0) as total_tokens,
+      COUNT(*)::integer as total_queries,
+      COALESCE(SUM(total_tokens)::integer, 0) as total_tokens,
       COALESCE(AVG(latency_ms), 0) as avg_latency_ms,
       COALESCE(SUM(CASE WHEN status = 'error' THEN 1 ELSE 0 END) * 100.0 / NULLIF(COUNT(*), 0), 0) as error_rate
     FROM llm_traces
@@ -85,7 +85,7 @@ export async function getLlmStats(hoursBack: number = 24): Promise<LlmStats> {
   `, [`-${hoursBack}`]);
 
   const modelBreakdown = await db.query<{ model: string; count: number; tokens: number }>(`
-    SELECT model, COUNT(*) as count, COALESCE(SUM(total_tokens), 0) as tokens
+    SELECT model, COUNT(*)::integer as count, COALESCE(SUM(total_tokens)::integer, 0) as tokens
     FROM llm_traces
     WHERE created_at >= NOW() + (? || ' hours')::INTERVAL
     GROUP BY model
